Simplifiy implementation
Before this change our RingBuffer awaited the replaced promise. While
this may be convinient a return value of `null` can now mean we did not
replace an object, or the replaced promise resolved to `null`.

So now instead our ringbuffer returns the previous promise and the
caller has to await it.

This makes a returnvalue of `null` clearer.